+++ inc/conn.h
@@ -1,6 +1,6 @@
 #ifndef _CONN_H_INCLUDED
 #define _CONN_H_INCLUDED
-#include "linebuf.h"
+#include "buffer.h"
 
 struct irc_conn {
     int fd;
@@ -9,8 +9,8 @@ struct irc_conn {
 
 struct irc_conn *irc_conn_open(const char *host, unsigned short port);
 int irc_conn_do_io(struct irc_conn *conn);
-int irc_conn_send(struct irc_conn *conn, const char *s);
-int irc_conn_sendf(struct irc_conn *conn, const char *fmt, ...);
+ssize_t irc_conn_send(struct irc_conn *conn, const char *s);
+ssize_t irc_conn_sendf(struct irc_conn *conn, const char *fmt, ...);
 void irc_conn_close_free(struct irc_conn *conn);
 
 #endif
